requirements.txt
langgraph==0.0.26
langchain==0.1.11
pandas==2.2.0
matplotlib==3.8.3
fpdf
pytrends==4.9.2
google-genai==0.3.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
fastapi==0.110.0
uvicorn==0.29.0
tavily-python==0.3.1
yfinance
python-pptx==0.6.23
requests==2.31.0